api/requirements.txt
fastapi
pydantic
uvicorn
numpy
typing
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
pillow
networkx==3.1
opencv-python-headless
tqdm
timm
prettytable
scipy
scikit-image
kornia